show number of connections in request logging

+++ src/main/java/edu/montana/csci/csci440/model/Album.java
@@ -74,7 +74,7 @@ public class Album extends Model {
         }
     }
 
-    public static Album find(int i) {
+    public static Album find(long i) {
         try (Connection conn = DB.connect();
              PreparedStatement stmt = conn.prepareStatement("SELECT * FROM albums WHERE AlbumId=?")) {
             stmt.setLong(1, i);
